backend/src/scraper/requirements.txt
flask>=3.0.0
playwright>=1.44.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0